Create 35. Check if Strings are rotations to each other or not
You are given two strings of equal lengths, s1 and s2. The task is to check if s2 is a rotated version of the string s1.

In C++, std::string::npos is a static member constant of the std::string class. It is used to indicate that a search operation (such as find, rfind, find_first_of, etc.) did not find the requested substring or character. The value of std::string::npos is typically the largest possible value for the size_t type, which is used as the return type for these search operations.

When a search function does not find the specified substring or character, it returns std::string::npos. This allows you to check whether the search was successful.